backend/requirements.txt
asgiref==3.3.1
django==3.1.4
django-environ==0.4.5
django-tailwind==1.1.0
gunicorn==20.0.4
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
django-cors-headers
djangorestframework==3.11.2